gallium/pipebuffer: change pb_cache_manager_create() size_factor to float
Requested by Marek.

Cc: "10.1" <mesa-stable@lists.freedesktop.org>

+++ src/gallium/auxiliary/pipebuffer/pb_bufmgr.h
@@ -161,8 +161,8 @@ pb_slab_range_manager_create(struct pb_manager *provider,
  */
 struct pb_manager *
 pb_cache_manager_create(struct pb_manager *provider, 
-                     	unsigned usecs,
-			unsigned size_factor,
+                        unsigned usecs,
+                        float size_factor,
                         unsigned bypass_usage);
